refactor(layering): Phase 8 A6 — path domain vocabulary into clm.core (refs #802)
Split clm.infrastructure.utils.path_utils: the course-domain vocabulary
(Lang/Format/Kind, OutputSpec, output_specs(), output_path_for(), ext_for,
skip/ignore rules, slide file-family detection, image dir constants,
prog-lang/extension mapping, is_in_dir, relative_path_to_course_img) moves
to a new clm.core.utils.path_utils; infrastructure keeps only the
filesystem helpers (find_project_root, atomic_write_all,
atomic_write_bytes). 43+ importers retargeted; the domain test file moves
to tests/core/utils/ with the atomic-write tests split back out.

Ratchet: clears the 11 core files whose only infrastructure import was
path_utils (42 -> 31 edges, 33 -> 23 files).
