+++ src/clm/cli/file_event_handler.py
@@ -6,8 +6,8 @@ from pathlib import Path
 from watchdog.events import PatternMatchingEventHandler
 
 from clm.core.course import Course
+from clm.core.utils.path_utils import is_ignored_dir_for_course
 from clm.infrastructure.backend import Backend
-from clm.infrastructure.utils.path_utils import is_ignored_dir_for_course
 
 logger = logging.getLogger(__name__)
