+++ src/clm/core/course_files/data_file.py
@@ -5,8 +5,8 @@ from attrs import define
 
 from clm.core.course_file import CourseFile
 from clm.core.utils.execution_utils import LAST_EXECUTION_STAGE
+from clm.core.utils.path_utils import is_ignored_file_for_output, output_specs
 from clm.infrastructure.operation import Concurrently, NoOperation, Operation
-from clm.infrastructure.utils.path_utils import is_ignored_file_for_output, output_specs
 
 if TYPE_CHECKING:
     from clm.core.output_target import OutputTarget
