+++ src/clm/core/operations/process_notebook.py
@@ -10,6 +10,13 @@ from typing import Any
 from attrs import frozen
 
 from clm.core.course_files.notebook_file import NotebookFile
+from clm.core.utils.path_utils import (
+    is_ignored_file_for_output,
+    is_image_file,
+    is_image_source_file,
+    output_path_for,
+    relative_path_to_course_img,
+)
 from clm.infrastructure.build_profiling import now as profiler_now
 from clm.infrastructure.build_profiling import profiler
 from clm.infrastructure.messaging.correlation_ids import (
@@ -18,13 +25,6 @@ from clm.infrastructure.messaging.correlation_ids import (
 )
 from clm.infrastructure.messaging.notebook_classes import NotebookPayload
 from clm.infrastructure.operation import Operation
-from clm.infrastructure.utils.path_utils import (
-    is_ignored_file_for_output,
-    is_image_file,
-    is_image_source_file,
-    output_path_for,
-    relative_path_to_course_img,
-)
 
 # Re-exported for existing callers/tests; the canonical home is the
 # infrastructure module so the diagram operations and the build's
