+++ tests/core/course_files/data_file_test.py
@@ -6,9 +6,9 @@ import pytest
 from clm.core.course_file import CourseFile
 from clm.core.course_files.data_file import DataFile
 from clm.core.operations.copy_file import CopyFileOperation
+from clm.core.utils.path_utils import output_specs
 from clm.infrastructure.backends.dummy_backend import DummyBackend
 from clm.infrastructure.operation import Concurrently, NoOperation
-from clm.infrastructure.utils.path_utils import output_specs
 
 DATA_FILE = "data/test.data"
