+++ tests/slides/test_voiceover_comment_token.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from clm.infrastructure.utils.path_utils import is_ignored_file_for_output
+from clm.core.utils.path_utils import is_ignored_file_for_output
 from clm.notebooks.slide_writer import format_narrative_cell
 from clm.slides.voiceover_tools import merge_voiceover_text, render_companion_update
